requirements.txt
numpy<2
torch==2.2.2
gymnasium==0.29.1
box2d==2.3.10
pygame==2.6.1
matplotlib
seaborn
imageio
moviepy
pandas
hydra-core>=1.3.2
omegaconf>=2.3.0
